moved streamlit_app.py to outside the web-app
